Any help? Alguma ajuda?
It is supposed to pic the letters that you put in and if it is an 'a',
replace it with a  '4'. Saving the progress in files. But...

É um programa pra pegar o que você digita e substituir todos os 'a' por
'4', salvando o progressoem arquivos. Mas...

The output file don't get any changes. (entrada, 'pyEntrada.mcod');

O arquivo de saída não é alterado de modo algum!